Move extract machinery to a file
