+++ flatgfa/src/ops/mod.rs
@@ -1,2 +1,3 @@
 pub mod depth;
+pub mod extract;
 pub mod position;
